Rename generator directory to interface-generator
