+++ src/core/converter/ApiDocToInterfaceConverter.ts
@@ -1,5 +1,5 @@
 import {endpointHasExamples, IApiDocEndpoint} from "../ApiDocInterfaces";
-import {InterfaceGenerator} from "../generator/InterfaceGenerator";
+import {InterfaceGenerator} from "../interface-generator/InterfaceGenerator";
 import {ApiDocEndpointParser} from "../parser/ApiDocEndpointParser";
 import {ApiDocExamplesParser} from "../parser/ApiDocExamplesParser";
