+++ src/core/interface-generator/InterfaceGenerator.test.ts
